+++ src/allotropy/parsers/agilent_gen5/agilent_gen5_structure.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import time
 from collections import defaultdict
 from dataclasses import dataclass
 from io import StringIO
@@ -8,7 +7,6 @@ import re
 
 import pandas as pd
 
-from allotropy.allotrope.models.shared.definitions.definitions import JsonFloat
 from allotropy.allotrope.models.shared.definitions.units import UNITLESS
 from allotropy.allotrope.schema_mappers.adm.plate_reader.rec._2024._06.plate_reader import (
     CalculatedDataItem,
@@ -17,6 +15,7 @@ from allotropy.allotrope.schema_mappers.adm.plate_reader.rec._2024._06.plate_rea
     MeasurementGroup,
     MeasurementType,
     Metadata,
+    ScanPositionSettingPlateReader,
 )
 from allotropy.exceptions import (
     AllotropeConversionError,
@@ -24,6 +23,7 @@ from allotropy.exceptions import (
 )
 from allotropy.parsers.agilent_gen5.constants import (
     ALPHALISA_FLUORESCENCE_FOUND,
+    DATA_SOURCE_FEATURE_VALUES,
     DEFAULT_SOFTWARE_NAME,
     DEVICE_TYPE,
     EMISSION_KEY,
@@ -42,7 +42,7 @@ from allotropy.parsers.agilent_gen5.constants import (
     ReadType,
     UNSUPPORTED_READ_MODE_ERROR,
     UNSUPPORTED_READ_TYPE_ERROR,
-    WAVELENGTHS_KEY, ScanPositionSettingPlateReader,
+    WAVELENGTHS_KEY,
 )
 from allotropy.parsers.constants import NOT_APPLICABLE
 from allotropy.parsers.lines_reader import SectionLinesReader
@@ -50,8 +50,8 @@ from allotropy.parsers.utils.pandas import read_csv, SeriesData
 from allotropy.parsers.utils.uuids import random_uuid_str
 from allotropy.parsers.utils.values import (
     try_float,
-    try_float_or_nan,
-    try_float_or_none, try_non_nan_float_or_none, try_non_nan_float,
+    try_float_or_none,
+    try_non_nan_float_or_none,
 )
 
 
@@ -91,28 +91,28 @@ class FilterSet:
     optics: str | None = None
 
     @property
-    def detector_wavelength_setting(self) -> JsonFloat | None:
+    def detector_wavelength_setting(self) -> float | None:
         if not self.emission:
             return None
-        return try_float_or_nan(self.emission.split("/")[0])
+        return try_non_nan_float_or_none(self.emission.split("/")[0])
 
     @property
-    def detector_bandwidth_setting(self) -> JsonFloat | None:
+    def detector_bandwidth_setting(self) -> float | None:
         if not self.emission:
             return None
         try:
-            return try_float_or_nan(self.emission.split("/")[1])
+            return try_non_nan_float_or_none(self.emission.split("/")[1])
         except IndexError:
             return None
 
     @property
-    def excitation_wavelength_setting(self) -> JsonFloat | None:
+    def excitation_wavelength_setting(self) -> float | None:
         if self.excitation:
-            return try_float_or_nan(self.excitation.split("/")[0])
+            return try_non_nan_float_or_none(self.excitation.split("/")[0])
         return None
 
     @property
-    def excitation_bandwidth_setting(self) -> JsonFloat | None:
+    def excitation_bandwidth_setting(self) -> float | None:
         if not self.excitation:
             return None
         try:
@@ -224,23 +224,25 @@ class KineticData:
     run_time: str | None
     interval: str | None
     reads: int | None
+    VALUES: int = 3
 
     @classmethod
     def create(cls, lines: list[str]) -> KineticData | None:
         for line in lines:
             if line.startswith("Start Kinetic\t"):
                 return cls._parse_kinetic_section(line)
+        return None
 
     @classmethod
     def _parse_kinetic_section(cls, line: str) -> KineticData:
         split_line = line.strip().split("\t")[1].split(", ")
-        if len(split_line) != 3:
+        if len(split_line) != KineticData.VALUES:
             msg = f"Expected the Kinetic section '{split_line}' to contain exactly 3 values."
             raise AllotropeConversionError(msg)
         return KineticData(
             run_time=split_line[0].split(" ")[1],
             interval=split_line[1].split(" ")[1],
-            reads=int(split_line[2].split(" ")[0])
+            reads=int(split_line[2].split(" ")[0]),
         )
 
 
@@ -384,7 +386,7 @@ class ReadData:
 
     @classmethod
     def _get_absorbance_measurement_labels(
-            cls, label_prefix: str | None, device_control_data: DeviceControlData
+        cls, label_prefix: str | None, device_control_data: DeviceControlData
     ) -> list[str]:
         pathlength_correction = device_control_data.get(PATHLENGTH_CORRECTION_KEY)
         measurement_labels = []
@@ -455,7 +457,7 @@ def get_identifiers(layout_lines: list[str] | None) -> dict[str, str]:
             well_pos = f"{row_name}{col_index + 1}"
             # Prefer Name to Well ID
             if not pd.isna(col) and (
-                    label == "Name" or label == "Well ID" and well_pos not in identifiers
+                label == "Name" or label == "Well ID" and well_pos not in identifiers
             ):
                 identifiers[well_pos] = col
     return identifiers
@@ -473,34 +475,38 @@ def get_temperature(actual_temperature_lines: list[str] | None) -> float | None:
     )
 
 
-def get_kinetic_measurements(kinetic_lines: list[str]) -> dict[str, list[str]] | None:
+def get_kinetic_measurements(
+    kinetic_lines: list[str] | None,
+) -> tuple[dict[str, list[float | None]], list[float]] | None:
     if not kinetic_lines:
         return None
     data = read_csv(StringIO("\n".join(kinetic_lines)), sep="\t")
-    kinetic_measurements = defaultdict(list)
+    kinetic_measurements: defaultdict[str, list[float | None]] = defaultdict(
+        list[float | None]
+    )
+    kinetic_elapsed_time: list[float] = []
     for col in data.columns:
-        for idx, value in enumerate(data[col]):
-            if not pd.isna(value):
-                kinetic_measurements[str(col)].append(str(value))
-
-    return {key: value for key, value in kinetic_measurements.items()}
+        for _idx, value in enumerate(data[col]):
+            if col == "Time":
+                kinetic_elapsed_time.append(_convert_time_to_seconds(str(value)))
+            elif not pd.isna(value):
+                kinetic_measurements[str(col)].append(float(str(value)))
+    return dict(kinetic_measurements.items()), kinetic_elapsed_time
 
 
 @dataclass(frozen=True)
 class MeasurementData:
     identifier: str
-    value: JsonFloat
+    value: float | None
     label: str
 
 
 def create_results(
-        result_lines: list[str],
-        header_data: HeaderData,
-        read_data: list[ReadData],
-        sample_identifiers: dict[str, str],
-        actual_temperature: float | None,
-        kinetic_data: KineticData | None,
-        kinetic_measurements: dict[str, list[str]] | None,
+    result_lines: list[str],
+    header_data: HeaderData,
+    read_data: list[ReadData],
+    sample_identifiers: dict[str, str],
+    actual_temperature: float | None,
 ) -> tuple[list[MeasurementGroup], list[CalculatedDataItem]]:
     if result_lines[0].strip() != "Results":
         msg = f"Expected the first line of the results section '{result_lines[0]}' to be 'Results'."
@@ -513,8 +519,8 @@ def create_results(
     well_to_measurements: defaultdict[str, list[MeasurementData]] = defaultdict(
         list[MeasurementData]
     )
-    calculated_data: defaultdict[str, list[tuple[str, JsonFloat]]] = defaultdict(
-        list[tuple[str, JsonFloat]]
+    calculated_data: defaultdict[str, list[tuple[str, float]]] = defaultdict(
+        list[tuple[str, float]]
     )
     measurement_labels = [
         label for r_data in read_data for label in r_data.measurement_labels
@@ -533,9 +539,6 @@ def create_results(
                 )
             else:
                 calculated_data[well_pos].append((label, well_value))
-
-    kinetic_elapsed_time = [_convert_time_to_seconds(time) for time in kinetic_measurements["Time"]] if kinetic_measurements else None
-    del kinetic_measurements["Time"]
 
     groups = [
         MeasurementGroup(
@@ -558,36 +561,103 @@ def create_results(
         for well_position, measurements in well_to_measurements.items()
     ]
 
-    if kinetic_measurements:
-        groups.extend(
-            MeasurementGroup(
-                measurement_time=header_data.datetime,
-                plate_well_count=len(well_to_measurements),
-                analytical_method_identifier=header_data.protocol_file_path,
-                experimental_data_identifier=header_data.experiment_file_path,
-                measurements=[
-                    _create_measurement(
-                        measurement := MeasurementData(random_uuid_str(), well_position, measurement_labels[0]),
-                        well_position,
-                        header_data,
-                        get_read_data_from_measurement(measurement, read_data),
-                        sample_identifiers.get(well_position),
-                        actual_temperature,
-                        kinetic_data,
-                        kinetic_measurements_data,
-                        kinetic_elapsed_time
-                    )
-                ],
-            ) for well_position, kinetic_measurements_data in kinetic_measurements.items() if well_position == "A1"
+    calculated_data_items = [
+        CalculatedDataItem(
+            identifier=random_uuid_str(),
+            data_sources=[
+                DataSource(
+                    identifier=measurement.identifier,
+                    feature=item.read_mode.value.lower(),
+                )
+                for measurement in _get_sources(
+                    label, well_to_measurements[well_position]
+                )
+                for item in read_data
+            ],
+            unit=UNITLESS,
+            name=label,
+            value=value,
         )
+        for well_position, well_calculated_data in calculated_data.items()
+        for label, value in well_calculated_data
+    ]
+
+    return groups, calculated_data_items
+
+
+def create_kinetic_results(
+    result_lines: list[str],
+    header_data: HeaderData,
+    read_data: list[ReadData],
+    sample_identifiers: dict[str, str],
+    actual_temperature: float | None,
+    kinetic_data: KineticData,
+    kinetic_measurements: dict[str, list[float | None]],
+    kinetic_elapsed_time: list[float],
+) -> tuple[list[MeasurementGroup], list[CalculatedDataItem]]:
+    if result_lines[0].strip() != "Results":
+        msg = f"Expected the first line of the results section '{result_lines[0]}' to be 'Results'."
+        raise AllotropeConversionError(msg)
+
+    # Create dataframe from tabular data and forward fill empty values in index
+    data = read_csv(StringIO("\n".join(result_lines[1:])), sep="\t")
+    data = data.set_index(data.index.to_series().ffill(axis="index").values)
+
+    calculated_data: defaultdict[str, list[tuple[str, float]]] = defaultdict(
+        list[tuple[str, float]]
+    )
+    measurement_labels = [
+        label for r_data in read_data for label in r_data.measurement_labels
+    ]
+    for row_name, row in data.iterrows():
+        label = row.iloc[-1]
+        for col_index, value in enumerate(row.iloc[:-1]):
+            well_pos = f"{row_name}{col_index + 1}"
+            well_value = try_non_nan_float_or_none(value)
+            # TODO: Report error documents for NaN values
+            if not well_value:
+                continue
+            calculated_data[well_pos].append((label, well_value))
+
+    groups = [
+        MeasurementGroup(
+            measurement_time=header_data.datetime,
+            plate_well_count=len(calculated_data),
+            analytical_method_identifier=header_data.protocol_file_path,
+            experimental_data_identifier=header_data.experiment_file_path,
+            measurements=[
+                _create_measurement(
+                    measurement := MeasurementData(
+                        random_uuid_str(), None, measurement_labels[0]
+                    ),
+                    well_position,
+                    header_data,
+                    get_read_data_from_measurement(measurement, read_data),
+                    sample_identifiers.get(well_position),
+                    actual_temperature,
+                    kinetic_data,
+                    kinetic_measurements[well_position],
+                    kinetic_elapsed_time,
+                )
+            ],
+        )
+        for well_position in calculated_data.keys()
+    ]
+
+    groups_by_well_position = {
+        group.measurements[0].location_identifier: group for group in groups
+    }
 
     calculated_data_items = [
         CalculatedDataItem(
-            identifier="UUID",
+            identifier=random_uuid_str(),
             data_sources=[
                 DataSource(
-                    identifier=random_uuid_str(),
-                    feature="absorption profile data cube",
+                    identifier=groups_by_well_position[well_position]
+                    .measurements[0]
+                    .identifier,
+                    # TODO: Add support for multiple kinetic sections
+                    feature=DATA_SOURCE_FEATURE_VALUES[read_data[0].read_mode],
                 )
             ],
             unit=UNITLESS,
@@ -602,10 +672,12 @@ def create_results(
 
 
 def get_read_data_from_measurement(
-        measurement: MeasurementData, read_data_list: list[ReadData]
+    measurement: MeasurementData, read_data_list: list[ReadData]
 ) -> ReadData:
     for read_data in read_data_list:
-        if _is_label_in_measurement_labels(measurement.label, read_data.measurement_labels):
+        if _is_label_in_measurement_labels(
+            measurement.label, read_data.measurement_labels
+        ):
             return read_data
 
     raise AllotropeConversionError(
@@ -614,7 +686,7 @@ def get_read_data_from_measurement(
 
 
 def _get_sources(
-        calculated_data_label: str, measurements: list[MeasurementData]
+    calculated_data_label: str, measurements: list[MeasurementData]
 ) -> list[MeasurementData]:
     # Pathlength is a special case, its sources are always determined
     # by the pathlength correction setting
@@ -645,11 +717,11 @@ def create_metadata(header_data: HeaderData) -> Metadata:
         software_version=header_data.software_version,
         file_name=header_data.file_name,
         asm_file_identifier=NOT_APPLICABLE,
-        data_system_instance_id=NOT_APPLICABLE
+        data_system_instance_id=NOT_APPLICABLE,
     )
 
 
-def _is_label_in_measurement_labels(label: str, measurement_labels: list[str]):
+def _is_label_in_measurement_labels(label: str, measurement_labels: set[str]) -> bool:
     if not measurement_labels:
         return False
     if label in measurement_labels:
@@ -658,18 +730,19 @@ def _is_label_in_measurement_labels(label: str, measurement_labels: list[str]):
     match = re.search(r"\[(\d+)]", label)
     if match and match.group()[1:-1] in measurement_labels:
         return True
+    return False
 
 
 def _create_measurement(
-        measurement: MeasurementData,
-        well_position: str,
-        header_data: HeaderData,
-        read_data: ReadData,
-        sample_identifier: str | None,
-        actual_temperature: float | None,
-        kinetic_data: KineticData | None = None,
-        kinetic_measurements: list[str] | None = None,
-        kinetic_elapsed_time: list[float] = None,
+    measurement: MeasurementData,
+    well_position: str,
+    header_data: HeaderData,
+    read_data: ReadData,
+    sample_identifier: str | None,
+    actual_temperature: float | None,
+    kinetic_data: KineticData | None = None,
+    kinetic_measurements: list[float | None] | None = None,
+    kinetic_elapsed_time: list[float] | None = None,
 ) -> Measurement:
     # TODO(switch-statement): use switch statement once Benchling can use 3.10 syntax
     if read_data.read_mode == ReadMode.ABSORBANCE:
@@ -681,22 +754,19 @@ def _create_measurement(
 
     if measurement_type is MeasurementType.ULTRAVIOLET_ABSORBANCE:
         filter_data = None
-        detector_wavelength_setting = float(
-            measurement.label.split(":")[-1].split(" ")[0]
-        )
+        detector_wavelength_setting = (
+            float(measurement.label.split(":")[-1].split(" ")[0])
+        ) or None
     else:
         filter_data = read_data.filter_sets[measurement.label]
         detector_wavelength_setting = filter_data.detector_wavelength_setting
-
-    if kinetic_measurements:
-        kinetic_measurements = [try_non_nan_float(value) for value in kinetic_measurements]
 
     return Measurement(
         type_=measurement_type,
         device_type=DEVICE_TYPE,
         identifier=measurement.identifier,
         sample_identifier=sample_identifier
-                          or f"{header_data.well_plate_identifier} {well_position}",
+        or f"{header_data.well_plate_identifier} {well_position}",
         location_identifier=well_position,
         well_plate_identifier=header_data.well_plate_identifier,
         detection_type=read_data.read_mode.value,
@@ -721,7 +791,8 @@ def _create_measurement(
         number_of_averages=read_data.number_of_averages,
         detector_carriage_speed=read_data.detector_carriage_speed,
         absorbance=measurement.value
-        if measurement_type == MeasurementType.ULTRAVIOLET_ABSORBANCE and not kinetic_data
+        if measurement_type == MeasurementType.ULTRAVIOLET_ABSORBANCE
+        and not kinetic_data
         else None,
         fluorescence=measurement.value
         if measurement_type == MeasurementType.FLUORESCENCE and not kinetic_data
@@ -730,19 +801,25 @@ def _create_measurement(
         if measurement_type == MeasurementType.LUMINESCENCE and not kinetic_data
         else None,
         compartment_temperature=actual_temperature,
-        total_measurement_time_setting=_convert_time_to_seconds(kinetic_data.run_time) if kinetic_data else None,
-        read_interval_setting=_convert_time_to_seconds(kinetic_data.interval) if kinetic_data else None,
+        total_measurement_time_setting=_convert_time_to_seconds(kinetic_data.run_time)
+        if kinetic_data
+        else None,
+        read_interval_setting=_convert_time_to_seconds(kinetic_data.interval)
+        if kinetic_data
+        else None,
         number_of_scans_setting=kinetic_data.reads if kinetic_data else None,
         dimensions_elapsed_time=kinetic_elapsed_time if kinetic_elapsed_time else None,
         measures=kinetic_measurements if kinetic_measurements else None,
     )
 
 
-def _convert_time_to_seconds(time_str: str) -> float:
+def _convert_time_to_seconds(time_str: str | None) -> float:
+    if not time_str:
+        return 0.0
     try:
         hours, minutes, seconds = map(float, time_str.split(":"))
         return hours * 3600 + minutes * 60 + seconds
 
     except ValueError:
         msg = f"Invalid time string: '{time_str}'."
-        raise AllotropeConversionError(msg)
+        raise AllotropeConversionError(msg) from None